GCG-26 moved domain objects to separate package
