--- codegen/src/main/java/org/hkust/jsonutils/Value.java
@@ -1,4 +1,0 @@
-package org.hkust.jsonutils;
-
-public interface Value {
-}
