+++ codegen/src/main/java/org/hkust/objects/Operator.java
@@ -1,4 +1,4 @@
-package org.hkust.jsonutils;
+package org.hkust.objects;
 
 public enum Operator {
     GREATER_THAN(">"),
