+++ codegen/src/main/java/org/hkust/objects/SelectCondition.java
@@ -1,4 +1,4 @@
-package org.hkust.jsonutils;
+package org.hkust.objects;
 
 import static java.util.Objects.requireNonNull;
