+++ codegen/src/main/java/org/hkust/objects/Value.java
@@ -1,0 +1,4 @@
+package org.hkust.objects;
+
+public interface Value {
+}
